+++ database/migrations/2024_05_30_050013_create_videos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('videos', function (Blueprint $table) {
             $table->id();
             $table->date('tanggal');
-            $table->integer('siswa_id');
+            $table->bigInteger('siswa_id');
             $table->string('video');
             $table->timestamps();
         });
